# public/openapi.json
{"openapi":"3.0.0","info":{"title":"MotorCato MESS API","version":"0.0.0"},"tags":[],"paths":{"/manufacturers":{"get":{"operationId":"Manufacturers_list","parameters":[],"responses":{"200":{"description":"The request has succeeded.","content":{"application/json":{"schema":{"type":"object","required":["jsonrpc","id","result"],"properties":{"jsonrpc":{"type":"string","enum":["2.0"]},"id":{"anyOf":[{"type":"integer","format":"int32"},{"type":"string"}],"nullable":true},"result":{"$ref":"#/components/schemas/ManufacturersResponse"}}}}}}}}},"/motor":{"get":{"operationId":"Motor_getReports","parameters":[{"name":"manufacturer","in":"query","required":true,"schema":{"type":"string"},"explode":false},{"name":"motor","in":"query","required":true,"schema":{"type":"string"},"explode":false}],"responses":{"200":{"description":"The request has succeeded.","content":{"application/json":{"schema":{"type":"object","required":["jsonrpc","id","result"],"properties":{"jsonrpc":{"type":"string","enum":["2.0"]},"id":{"anyOf":[{"type":"integer","format":"int32"},{"type":"string"}],"nullable":true},"result":{"$ref":"#/components/schemas/MotorSummary"}}}}}}}}},"/motors":{"get":{"operationId":"Motors_listByBrand","parameters":[{"name":"manufacturer","in":"query","required":true,"schema":{"type":"string"},"explode":false}],"responses":{"200":{"description":"The request has succeeded.","content":{"application/json":{"schema":{"type":"object","required":["jsonrpc","id","result"],"properties":{"jsonrpc":{"type":"string","enum":["2.0"]},"id":{"anyOf":[{"type":"integer","format":"int32"},{"type":"string"}],"nullable":true},"result":{"$ref":"#/components/schemas/MotorsResponse"}}}}}}}}}},"components":{"schemas":{"Manufacturer":{"type":"object","required":["name","failureCount","searchURL","source_url"],"properties":{"name":{"type":"string","description":"The official name of the manufacturer (e.g., 'AeroTech')"},"failureCount":{"type":"integer","format":"int32","description":"Total number of failures across all motors for this manufacturer"},"lastFailure":{"type":"string","description":"The most recent failure data for the manufacturer."},"searchURL":{"type":"string","enum":["/search?manufacturer=Estes%20Industries"],"description":"Path for a search of all failures for the manufacturer."},"source_url":{"type":"string","format":"uri","description":"Link to a search of all failures for the manufacturer on MotorCato."}},"example":{"name":"Estes Industries","failureCount":1204,"lastFailure":"2025-06-14","searchURL":"/search?manufacturer=Estes%20Industries","source_url":"https://www.motorcato.org/search?manufacturer=Estes%20Industries"}},"ManufacturersResponse":{"type":"object","required":["count","manufacturers","source_url"],"properties":{"count":{"type":"integer","format":"int32","description":"The number of manufacturers known."},"manufacturers":{"type":"array","items":{"$ref":"#/components/schemas/Manufacturer"},"description":"The known manufacturers."},"source_url":{"type":"string","format":"uri","description":"Link to MotorCato."}}},"MotorSummary":{"type":"object","required":["manufacturer","motor","failureCount","searchURL","source_url"],"properties":{"manufacturer":{"type":"string","description":"The manufacturer name"},"motor":{"type":"string","description":"The motor common name (e.g., 'G25')"},"failureCount":{"type":"integer","format":"int32","description":"Number of recorded failure reports for this motor."},"lastFailure":{"type":"string","description":"The most recent failure data for the motor."},"searchURL":{"type":"string","description":"Path for a search of all failures for the motor."},"source_url":{"type":"string","format":"uri","description":"Link to a search of all failures for the motor on MotorCato."}},"example":{"manufacturer":"AeroTech","motor":"G25","failureCount":3,"lastFailure":"2020-11-07","searchURL":"/search?manufacturer=AeroTech&name=G25","source_url":"https://www.motorcato.org/search?manufacturer=AeroTech&name=G25"}},"MotorsResponse":{"type":"object","required":["count","motors","source_url"],"properties":{"count":{"type":"integer","format":"int32","description":"The number of motors known for this manufacturer."},"motors":{"type":"array","items":{"$ref":"#/components/schemas/MotorSummary"},"description":"The known motors for this manufacturer."},"source_url":{"type":"string","format":"uri","description":"Link to a search of all failures for the manufacturer on MotorCato."}}}}},"servers":[{"url":"https://www.motorcato.org/api/1","description":"Primary Production Server","variables":{}}]}